Update results.qmd including new figures
This commit modifies the results.qmd file to improve the presentation of biodiversity metrics across elevational gradients. Key changes include the addition of a self-contained HTML format, updates to figure labels and dimensions, and the introduction of new plots for bioclimatic correlations and PCA analyses. The restructuring of sections enhances clarity and aligns with the overall project objectives.